requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
pyee==7.0.2
slackclient==2.7.3
slackeventsapi==2.2.1
typing-extensions==3.7.4.2
Werkzeug==1.0.1
yarl==1.5.0
